LayoutTests/js/dom/Promise-then-in-workers-expected.txt

[Worker] Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../resources/Promise-then-in-workers.js
PASS [Worker] thisInInit === firstPromise is false
PASS [Worker] thisInInit === global is true
PASS [Worker] firstPromise instanceof Promise is true
PASS [Worker] secondPromise instanceof Promise is true
PASS [Worker] thisInFulfillCallback === firstPromise is false
PASS [Worker] thisInFulfillCallback === secondPromise is false
PASS [Worker] thisInFulfillCallback === global is true
PASS [Worker] result is "hello"
PASS [Worker] result is "world"
PASS [Worker] rejected
PASS [Worker] result is "exception"
PASS [Worker] resolved
PASS successfullyParsed is true

TEST COMPLETE
